# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Task",
        "description": "Auth with JWT, and users can create Statements",
        "contact": {
            "email": "ashotshushanyan14@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/statements/{id}": {
            "put": {
                "tags": [
                    "Statement"
                ],
                "summary": "Update an existing statement",
                "description": "Update an existing statement and its related files.",
                "operationId": "2a0b3a9cbbbd0694d58c733dbe6dbf6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the statement to update.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The data needed to update the statement.",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The statement was updated successfully."
                    },
                    "404": {
                        "description": "The statement with the provided ID was not found."
                    },
                    "422": {
                        "description": "The data provided is invalid or incomplete."
                    },
                    "500": {
                        "description": "An error occurred while processing the request."
                    }
                }
            }
        },
        "/api/statement-files": {
            "post": {
                "tags": [
                    "StatementFile"
                ],
                "summary": "Store a new StatementFile",
                "description": "Store a new StatementFile with the provided data",
                "operationId": "storeStatementFile",
                "requestBody": {
                    "description": "Provide the necessary fields to create a new StatementFile",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file",
                                    "filename",
                                    "statement_id"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "base54"
                                    },
                                    "filename": {
                                        "type": "string",
                                        "example": "file.pdf"
                                    },
                                    "statement_id": {
                                        "type": "integer",
                                        "example": "{statement_id}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "StatementFile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "filename": {
                                            "type": "string",
                                            "example": "{filename}"
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "example": "storage/app/statements/{filename}"
                                        },
                                        "statement_id": {
                                            "type": "integer",
                                            "example": "{statement_id}"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "{created_at}"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "{updated_at}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "file": [
                                                    "The file field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error occurred while storing the StatementFile."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/statement-files/{id}": {
            "put": {
                "tags": [
                    "StatementFile"
                ],
                "summary": "Update an existing StatementFile",
                "description": "Update an existing StatementFile with the provided data",
                "operationId": "updateStatementFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the StatementFile to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide the fields to update a StatementFile",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "file_name": {
                                        "type": "string",
                                        "example": "updated_file.pdf"
                                    },
                                    "file_path": {
                                        "type": "string",
                                        "example": "updated_file_path"
                                    },
                                    "statement_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "StatementFile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "file_name": {
                                            "type": "string",
                                            "example": "updated_file.pdf"
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "example": "updated_file_path"
                                        },
                                        "statement_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "{created_at}"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "{updated_at}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "StatementFile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The specified StatementFile could not be found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "file_name": [
                                                    "The file name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error occurred while updating the StatementFile."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "StatementFile"
                ],
                "summary": "Delete a StatementFile",
                "description": "Delete a StatementFile with the specified ID",
                "operationId": "deleteStatementFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the StatementFile to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "StatementFile deleted successfully"
                    },
                    "404": {
                        "description": "StatementFile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The specified StatementFile could not be found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server error occurred while deleting the StatementFile."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}